refactor(web): web の repository thin wrapper を単一エントリに集約する
shared/api/repository を web の client で合成する処理が 8 ファイルに分散
していたのを web/src/repository.ts 1 つにまとめる。
合成・export 名は同一なので call site の名前 import は変えない、
パスだけ ~/repository/xxx-repository → ~/repository に揃える。
