+++ web/src/hooks/useDashboard.ts
@@ -5,47 +5,39 @@ import {
   useQueryClient,
 } from "@tanstack/react-query";
 import {
+  closePoll,
+  convertFile,
   createBroadcast,
-  deleteBroadcast,
-  fetchBroadcasts,
-  sendBroadcastNow,
-  updateBroadcast,
-  uploadBroadcastImage,
-} from "~/repository/broadcast-repository";
-import { fetchEmergencies } from "~/repository/emergency-repository";
-import {
+  createPoll,
   deleteAllFeedbacks,
+  deleteAllKnowledge,
+  deleteAllPersonas,
+  deleteBroadcast,
+  deleteFile,
+  deletePoll,
+  extractPersonas,
+  fetchBroadcasts,
+  fetchEmergencies,
   fetchFeedbackById,
   fetchFeedbacks,
-  resolveFeedback,
-  unresolveFeedback,
-} from "~/repository/feedback-repository";
-import {
-  convertFile,
-  deleteAllKnowledge,
-  deleteFile,
   fetchFileContent,
   fetchFiles,
-  fetchUnifiedFiles,
-  reconvertFile,
-  saveFile,
-  syncKnowledge,
-  uploadFile,
-} from "~/repository/knowledge-repository";
-import {
-  deleteAllPersonas,
-  extractPersonas,
   fetchPersonas,
-} from "~/repository/persona-repository";
-import {
-  closePoll,
-  createPoll,
-  deletePoll,
   fetchPollResultsAdmin,
   fetchPolls,
+  fetchUnifiedFiles,
+  reconvertFile,
+  resolveFeedback,
+  saveFile,
+  sendBroadcastNow,
   sendPollNow,
+  syncKnowledge,
+  unresolveFeedback,
+  updateBroadcast,
   updatePoll,
-} from "~/repository/poll-repository";
+  uploadBroadcastImage,
+  uploadFile,
+} from "~/repository";
 import type { PollStatus } from "~/types";
 
 export const dashboardKeys = {
