--- web/src/repository/index.ts
@@ -1,7 +1,0 @@
-export * from "./broadcast-repository";
-export * from "./emergency-repository";
-export * from "./feedback-repository";
-export * from "./invitation-repository";
-export * from "./knowledge-repository";
-export * from "./persona-repository";
-export * from "./thread-repository";
